The deps file seems to be working with these changes
The deps file is built differently now.
The js subfolders need a link to closure-library and pearson in order for the deps file to work, and the goog/base.js much be included as:
<script src="js/closure-library/closure/goog/base.js"></script>

I added a script which can create the 5 needed symlinks when executed from the js folder where the links should be put.